compilador_python/requirements.txt
Pillow>=10.0.0,<11.0.0
reportlab>=4.0.0
ttkbootstrap>=1.10.1
tkinterdnd2>=0.3.0
nltk>=3.8.1
mss>=9.0.1
pyscreenshot>=3.1
fpdf2>=2.7.8
darkdetect>=0.7.1
sv-ttk>=2.5.5
